requirements.txt
requests
python-dotenv
sqlalchemy
pandas
python-telegram-bot
numpy
streamlit
plotly
redis